requirements.txt
streamlit
openai
pyvis
pdfplumber
python-dotenv